LayoutTests/fast/profiler/document-dot-write-expected.txt

CONSOLE MESSAGE: line 59: Call Document.write()
CONSOLE MESSAGE: line 75: Thread_1  0
CONSOLE MESSAGE: line 75:  (program) document-dot-write.html 8
CONSOLE MESSAGE: line 75:   write  0
CONSOLE MESSAGE: line 75:   endTest profiler-test-JS-resources.js 1
CONSOLE MESSAGE: line 75:  (idle)  0
This was a triumph. This page calls document.write(). 

To run this test manually, load it in the browser then load the WebInspector and look at the profile. If there is a profile it should show write().
